database_sql_jdbc_template/src/test/java/com/miguel/database_sql/dao/Impl/BookDaoImplTest.java
package com.miguel.database_sql.dao.Impl;

import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.verify;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentMatchers;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.jdbc.core.JdbcTemplate;

import com.miguel.database_sql.TestDataUtil;
import com.miguel.database_sql.dao.impl.BookDaoImpl;
import com.miguel.database_sql.domain.Book;

@ExtendWith(MockitoExtension.class)
public class BookDaoImplTest {

    @Mock
    private JdbcTemplate jdbcTemplate;

    @InjectMocks
    private BookDaoImpl underTest;

    @Test
    public void testThatCreateBookGenereatesCorrectSQL() {
        Book book = TestDataUtil.createTestBook();

        underTest.create(book);

        verify(jdbcTemplate).update(
                eq("INSERT INTO books (isbn, title, author_id) VALUES (?, ?, ?)"),
                eq("12d2131d"),
                eq("Book Title"),
                eq(1L));
    }

    @Test
    public void testThatFindOneBookGeneratesTheCreateSQL() {
        underTest.findOne("12d2131d");

        verify(jdbcTemplate).query(eq("SELECT isbn, title, author_id FROM books WHERE isbn = ? LIMIT 1"),
                ArgumentMatchers.<BookDaoImpl.BookRowWrapper>any(),
                eq("12d2131d"));
    }

}
